+++ packages/iov-ethereum/types/parse.d.ts
@@ -1,4 +1,5 @@
-import { Amount } from "@iov/bcp";
+import { Amount, TransactionId } from "@iov/bcp";
 export declare class Parse {
     static ethereumAmount(total: string): Amount;
+    static transactionId(hash: string): TransactionId;
 }